refactor: Lotto 클래스의 패키지 변경
Lotto 클래스를 model 패키지 하위로 이동
getter 및 로또 번호가 존재하는지 확인하는 메서드 추가
