+++ src/main/java/lotto/model/Lotto.java
@@ -1,5 +1,6 @@
-package lotto;
+package lotto.model;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class Lotto {
@@ -17,4 +18,11 @@ public class Lotto {
     }
 
     // TODO: 추가 기능 구현
+    public List<Integer> getNumbers() {
+        return new ArrayList<>(numbers);
+    }
+
+    public boolean containsNumber(Integer number) {
+        return numbers.contains(number);
+    }
 }
